+++ agbenchmark/challenges/adapatability/a1_debug/artifacts_in/sample_code.py
