+++ agbenchmark/challenges/code/c1_writing_suite_1/1_return/artifacts_in/sample_code.py
