+++ agbenchmark/challenges/code/c1_writing_suite_1/1_return/artifacts_out/sample_code.py
