+++ agbenchmark/challenges/code/c1_writing_suite_1/2_write/artifacts_out/sample_code.py
