+++ agbenchmark/challenges/code/c1_writing_suite_1/3_modify/artifacts_in/sample_code.py
