+++ agbenchmark/challenges/code/c1_writing_suite_1/4_tests/artifacts_in/sample_code.py
