+++ agbenchmark/challenges/code/c1_writing_suite_1/4_tests/data.json
@@ -1,7 +1,7 @@
 {
   "name": "TestReturnCode_Tests",
   "category": ["code", "iterate"],
-  "task": "First, modify test.py to fill in the test case to be able to test the code in code.py. Next, modify the multiply_int function in code.py to be able to pass in a 'multiplier' argument to multiply the 'num' by 'multiplier'. Both arguments are integers. You can make sure you have correctly done this by running test.py that you previously modified.",
+  "task": "First, modify testfile.py to fill in the test case to be able to test the code in sample_code.py. Next, modify the multiply_int function in sample_code.py to be able to pass in a 'multiplier' argument to multiply the 'num' by 'multiplier'. Both arguments are integers. You can make sure you have correctly done this by running testfile.py that you previously modified.",
   "dependencies": ["TestReturnCode_Modify"],
   "cutoff": 120,
   "ground": {
