+++ agbenchmark/challenges/code/c2_debug_suite/d2.1_guided/artifacts_in/sample_code.py
