+++ agbenchmark/challenges/code/c2_debug_suite/d2.1_guided/artifacts_out/sample_code.py
