+++ agbenchmark/challenges/code/c2_debug_suite/d2.2_vague/artifacts_in/sample_code.py
