+++ agbenchmark/challenges/code/c2_debug_suite/d2.2_vague/artifacts_in/test.py
@@ -1,6 +1,7 @@
 # mypy: ignore-errors
-from code import two_sum
 from typing import List
+
+from sample_code import two_sum
 
 
 def test_two_sum(nums: List, target: int, expected_result: List[int]) -> None:
