+++ agbenchmark/challenges/code/c2_debug_suite/d2.2_vague/artifacts_out/sample_code.py
