+++ agbenchmark/challenges/code/c3_writing_suite_2/d3.1_three_sum/custom_python/test.py
@@ -1,6 +1,7 @@
 # mypy: ignore-errors
-from code import three_sum
 from typing import List
+
+from sample_code import three_sum
 
 
 def test_three_sum(nums: List[int], target: int, expected_result: List[int]) -> None:
