+++ agbenchmark/challenges/code/c3_writing_suite_2/d3_two_sum/artifacts_out/sample_code.py
