+++ agbenchmark/challenges/code/c4_writing_cli_suite_3/1_password_generator/data.json
@@ -1,7 +1,7 @@
 {
   "name": "TestPasswordGenerator_Easy",
   "category": ["code"],
-  "task": "Create a random password generator. The password should have between 8 and 16 characters and should contain letters, numbers and symbols. The password should be printed to the console. The entry point will be a python file that can be run this way: python password_generator.py",
+  "task": "Create a random password generator. The password should have between 8 and 16 characters and should contain letters, numbers and symbols. The password should be printed to the console. The entry point will be a python file that can be run this way: python password_generator.py [--len x] where x is the length of the password. If no length is specified, the password should be 8 characters long. The password_generator can also be imported as a module and called as password = password_generator.generate_password(len=x).",
   "dependencies": ["TestWriteFile"],
   "cutoff": 90,
   "ground": {
@@ -15,7 +15,7 @@
   },
   "info": {
     "difficulty": "basic",
-    "description": "Tests ability for the agent to code a file organizer.",
+    "description": "Tests ability for the agent to create a random password generator.",
     "side_effects": []
   }
 }
